+++ luscii_patient_actions_sdk_platform_interface/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.7.0] - 2026-01-20
+
+### Added
+- `LusciiEnvironment` enum to support Production, Acceptance, and Test environments.
+- `environment` parameter to `initialize` method.
 
 ## [0.6.0+1] - 2025-11-24